requirements.txt
# Gmail API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# Anthropic Claude API
anthropic==0.39.0

# Environment variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
